Unify the query format between menus and entry's
